refactor: refactor the AuthorizationOwnerType enum location
