+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/state/appliers/AuthorizationCreatedApplier.java
